fix

See merge request polytechnique-montr-al/log3900/26-1/equipe-105/LOG3900-105!54

+++ client_leger/lib/features/game_session/presentation/widgets/item_card/item_card_widget.dart
@@ -241,8 +241,8 @@ class ItemCardWidget extends StatelessWidget {
     final widgets = <Widget>[];
 
     for (final top in positions) {
-      widgets.add(_diamond(left: 6, top: top.toDouble()));
-      widgets.add(_diamond(right: 6, top: top.toDouble()));
+      widgets.add(_diamond(left: 6, top: top));
+      widgets.add(_diamond(right: 6, top: top));
     }
     return widgets;
   }
